separate observation space decorator classes & DQN training updates
