+++ robo_rugby/__init__.py
@@ -16,3 +16,11 @@ register(
     nondeterministic=True,  # opposing team will pretty much always be nondeterministic
     reward_threshold=1.0  # no idea
 )
+
+register(
+    id='RoboRugbySimpleDuel-v2',
+    entry_point='robo_rugby.gym_env.RR_Environments:SimpleDuel2',
+    max_episode_steps=const.GAME_LENGTH_STEPS,
+    nondeterministic=True,  # opposing team will pretty much always be nondeterministic
+    reward_threshold=1.0  # no idea
+)
