+++ robo_rugby/gym_env/RR_Constants.py
@@ -1,15 +1,15 @@
 import pygame
 import math
 
-ARENA_WIDTH = 800
-ARENA_HEIGHT = 800
+ARENA_WIDTH = 400  #800
+ARENA_HEIGHT = 400  #800
 ROBOT_LENGTH = 20
 ROBOT_WIDTH= 40
 ROBOT_WIDTH_BODY = 32
 ROBOT_WIDTH_TRACKS = 8
-ROBOT_VEL = 12 #3
+ROBOT_VEL = 12  #3
 MOVES_PER_FRAME = ROBOT_VEL
-ROBOT_ANGULAR_VEL_ONE = .9  #.6 # one motor rotating around other track
+ROBOT_ANGULAR_VEL_ONE = .9  # one motor rotating around other track
 ROBOT_ANGULAR_VEL_BOTH = 1.5  # with both motors engaged
 GOAL_WIDTH = 240
 GOAL_HEIGHT = 240
@@ -18,34 +18,34 @@ BALL_RADIUS = 7
 BALL_SLOWDOWN = .995
 BALL_MIN_SPEED = 0.005
 PUSH_FACTOR = BALL_SLOWDOWN
-FRAMERATE = 30 #120
-GAME_LENGTH_MINS = .15 #3
+FRAMERATE = 30  #120
+GAME_LENGTH_MINS = .15  #3
 GAME_LENGTH_STEPS = int(GAME_LENGTH_MINS * 60 * FRAMERATE)
 
 TIME_BALL_IN_GOAL_SECONDS = 5
 TIME_BALL_IN_GOAL_STEPS = TIME_BALL_IN_GOAL_SECONDS * FRAMERATE
 
-NUM_BALL_POS = 1
-NUM_BALL_NEG = 1
-NUM_ROBOTS_HAPPY = 1
-NUM_ROBOTS_GRUMPY = 1
+NUM_BALL_POS = 1  #4
+NUM_BALL_NEG = 0  #4
+NUM_ROBOTS_HAPPY = 1  #2
+NUM_ROBOTS_GRUMPY = 0  #2
 NUM_ROBOTS_TOTAL = NUM_ROBOTS_GRUMPY + NUM_ROBOTS_HAPPY
 
 MAX_NEG_BALLS = 3
 POINTS_BALL_SCORED = 500
-POINTS_TIME_PENALTY = .1 # Time penalty per frame
+POINTS_TIME_PENALTY = .1  # Time penalty per frame
 POINTS_ROBOT_CRASH_PENALTY = .005
 POINTS_ROBOT_IN_GOAL_PENALTY = .005
 POINTS_NO_MOVE_PENALTY = .005
 # Let's say that if a ball starts at (0,0) and travels all the way to (ARENA_WIDTH,ARENA_HEIGHT)
 # the ball would get this many points:
-POINTS_BALL_TRAVEL_MAX = 2000
+POINTS_BALL_TRAVEL_MAX = 200000
 # Then for each pixel moved closer to the goal it gets:
 POINTS_BALL_TRAVEL_MULT = POINTS_BALL_TRAVEL_MAX / math.pow(ARENA_WIDTH**2 + ARENA_HEIGHT**2, .5)
 # If you destroy a goal, get all the points you possibly could from dunking all the balls
 POINTS_GOAL_DESTROYED = (POINTS_BALL_SCORED + POINTS_BALL_TRAVEL_MAX) * (NUM_BALL_POS + NUM_BALL_NEG)
 
-POINTS_ROBOT_TRAVEL_MULT = POINTS_BALL_TRAVEL_MULT  # for now
+POINTS_ROBOT_TRAVEL_MULT = POINTS_BALL_TRAVEL_MULT / 100  # for now
 
 TEAM_HAPPY = 1
 TEAM_GRUMPY = -1
